ij-legacy: Avoid linking directly into ImageJ 1.x
It was a mistake to assume that we have direct access to all the ImageJ
1.x classes and to assume as a consequence that we could just import
them left and right and not care about clean separation between ImageJ
1.x classes and ImageJ 2.x classes.

Now this comes back to bite us: in the current form, we would never be
able to interact with more than one ImageJ 1.x 'context' encapsulated in
its own class loader.

So what we will have to do now is undo the damage and introduce the
clean separation (which is pretty painful by now).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>